Start of some bigger refactoring
Move GameSignals.h into the server from the client repository.
Made handlePacket use the Net::Signaling.
Started work on adding bots to the rooms.
Made static maps for StatsHolder class, with binding from strings to relevant
variables and linked it to the get/send data functions.
Changed all sf::Uint to the the equivalent uint_t and int_t.
Made a first try at a sendScores function for challenges to move them to
the database.